Refactor account management page
Now 50 users are displayed on each page, instead of infinite. Most of
the edit links are now gone, as you can now easily go to the profile
page. Explicit links to edit your own account have been removed (can
still be found in the table and on your profile page).