Link whole article teaser to article
The title, image, intro and of course the read more link are now linked
to the article. The title, image and intro no longer rely on JavaScript,
but still have a disguised link (no colored/underlined text). The css
class disguised_link, first used for comments, is used for this.