Move authentication to the request object
The current setup is not ideal; the explicit dependency on the user repository in hasRank and getCurrentUser is ugly. However, now at least the Website object is almost free of request state. (Only the getRequestString and getRequestInt methods remain.)

The rank logic was placed in the Authentication class. It's still there, but it would be better if it is in a separate class. This class could then be accessed from Website.