Prevent need for writing during theme switch
The writing made it practically impossible to run the website in the same directory as its source code was stored in. This makes development harder.

This new setup splits theme files across two locations: a web-inaccessible and a web-accessible one. This unfortunatily makes themes slightly harder to develop (no longer everything is stored in a single folder) and slightly harder to distribute. Would there be a better solution, while still having some theme files in a private folder, and others in a public folde?